<SideMenu> retracting at load fix.
This happened due to how the animations were handled, with the state
for tracking this being a boolean that started as false, adding the
'--hide' CSS class to all <SideMenu>s on load, causing them to animate.

The boolean state has been replaced with a string state (more readable
than a number with -1, 0 and 1 as values) to track three values, where
they start as neutral, with no CSS class being added.