<SideMenu> event listeners optimization.
A useRef to track the value of 'showSideMenu' was added to allow its
value to be read from inside those event listeners. Now, they'll only
run when the <SideMenu> is on 'show'.

This also makes so there's no need to check if the 'showSideMenu' is on
neutral inside 'setShowSideMenu()' at the end of those eventListeners.