<TextInput> buttons visiblity fix.
They were only disappearing if the clear button was used. This was due
to the button setting the "translate" useState to [''] and the validation
on whether or not to show the buttons was reliant on that, which isn't
the case when using Ctrl-Z or simply backspacing it all away.

It had been done like this because of the <input>'s "value" attribute on
<TextInput>, but then the value was dropped in favour of changing the
<input> via useRef because "translation.join('\n')" was messing up with
both Ctrl-Z and Ctrl-X.