<TextInput> textToTranslate splitting rework.
The previous .split() was meant to be a concise way to weed out
Templates from the rest of the article body, and it worked for
well-structured articles or single Templates. The problem was that it
couldn't handle articles with multiple back-to-back Templates.

Although not as concise, the new split will produce a better array of
strings more times than the previous version.

Other changes:
- New 'splitRawInput()' function to deal with splitting the input;

- New debug print for skipped parts of the article body;

- Done away with the 'handleHistoryUpdate()', since there's no need to
  handle all the {{UL}} anymore;

- Removed treatment for anything other than Templates, as those get
  weeded out during splitting;

- Changed the return values for 'translation()' to keep them consistent,
  so now every single line is returned individually and <br/> only gets
  added after 'formatLine()';

- Removed the bit of code responsible for keeping untranslated missing
  Template text with the equals symbol unmaked. It looked cool and all,
  but was too much spaghetti for so little return;

- Removed [null] returns.